+++ src/__tests__/Checkbox.test.jsx
@@ -3,18 +3,18 @@ import { test , expect } from "vitest";
 import { render, fireEvent, waitFor } from "@testing-library/react";
 import { Provider } from "react-redux";
 import { configureStore } from "@reduxjs/toolkit";
-import inputsSlice from "../../reducer/inputsSlice";
+import inputsSlice from "../reducer/inputsSlice";
 import { TouchBackend } from "react-dnd-touch-backend";
-import Checkbox from "../Checkbox";
+import {CheckBox} from "../components";
 import "@testing-library/jest-dom";
 import { DndProvider } from "react-dnd";
-test("Checkbox should render and handle change event", async () => {
+test("CheckBox should render and handle change event", async () => {
   const store = configureStore({ reducer: { inputs: inputsSlice } });
 
   const { getByRole } = render(
     <Provider store={store}>
       <DndProvider backend={TouchBackend}>
-        <Checkbox />
+        <CheckBox />
       </DndProvider>
     </Provider>
   );
