+++ src/__tests__/DropZone.test.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { test, expect } from "vitest";import { render } from "@testing-library/react";
-import DropZone from "../DropZone";
+import {DropZone} from "../components";
 import "@testing-library/jest-dom";
 test("DropZone should render correctly", () => {
   const headerText = "Test Header";
