+++ src/__tests__/DropZones.test.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { test, expect } from "vitest";import { render, screen, fireEvent } from "@testing-library/react";
-import DragZones from "../DragZones";
+import {DragZones} from "../components";
 import { Provider } from "react-redux";
-import { store } from "../../store/store";
+import { store } from "../store/store";
 import "@testing-library/jest-dom";
 import { DndProvider } from "react-dnd";
 import { TouchBackend } from "react-dnd-touch-backend";
