+++ src/__tests__/TextInput.test.jsx
@@ -1,11 +1,11 @@
 import React from "react";
 import { test, expect } from "vitest";import { render, fireEvent } from "@testing-library/react";
-import TextInput from "../TextInput";
+import {TextInput} from "../components";
 import "@testing-library/jest-dom";
 import { DndProvider } from "react-dnd";
 import { TouchBackend } from "react-dnd-touch-backend";
 import { Provider } from "react-redux";
-import { store } from "../../store/store";
+import { store } from "../store/store";
 test("TextInput should render and handle change event", async () => {
   const { getByPlaceholderText } = render(
     <Provider store={store}>
