+++ src/hooks/__tests__/useDragDrop.test.jsx
@@ -1,8 +1,10 @@
+import React from "react";
 import { renderHook } from "@testing-library/react-hooks";
 import { configureStore } from "@reduxjs/toolkit";
 import { Provider } from "react-redux";
 import rootReducer from "../../reducer/inputsSlice"; // Import your rootReducer here
 import useDragDrop from "../useDragDrop";
+import {  describe  , it } from "vitest";
 
 describe("useDragDrop", () => {
   it("should handle drag and drop events", () => {
