+++ src/hooks/useRenderZone.tsx
@@ -11,7 +11,7 @@ const useRenderZone = (
         (component: States.StoredComponentState) => component.id === id
       );
     const list = components?.map((component: States.ComponentState) =>
-      condition(component.id) ? (<div key={component.id}>{ component.el } </div>)  : null
+      condition(component.id) ? component.el   : null
     );
 
     return list;
