+++ src/main.tsx
@@ -13,4 +13,4 @@ ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
       </PersistGate>
     </Provider>
   </React.StrictMode>
-);
+);
